sam-backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pillow==10.2.0
numpy==1.26.3
torch>=2.0.0
torchvision>=0.15.0
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git
pydantic==2.5.3
supabase>=2.0.0
python-dotenv>=1.0.0
